+++ animal.cpp
@@ -1,0 +1,1 @@
+#include "animal.hpp"
